Resolves #95, adding a refresh button at MainWalletScreen
